corrección al buscar usuario en listado Ajustes en lata medidores del dashboard usuario

+++ src/main/java/com/hardnets/coop/model/dto/AllTariffsDto.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 
 @Getter
@@ -15,7 +14,7 @@ public class AllTariffsDto {
     private Long id;
     private String waterMeterSize;
     private Integer flatFee;
-    private BigDecimal cubicMeter;
+    private Float cubicMeter;
     private String clientType;
     private Instant lastUpdate;
