+++ src/main/java/com/hardnets/coop/model/dto/TariffDto.java
@@ -5,14 +5,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class TariffDto {
     private Long id;
-    private BigDecimal cubicMeter;
+    private Float cubicMeter;
     private Integer flatFee;
     private Long clientId;
     private Long sizeId;
